+++ R/rankinma-package.R
@@ -71,6 +71,7 @@
 #' @importFrom graphics text
 #' @importFrom graphics title
 #' @importFrom grDevices col2rgb
+#' @importFrom grDevices palette.colors
 #' @importFrom grDevices rainbow
 #' @importFrom grDevices recordPlot
 #' @importFrom grDevices rgb